+++ classes/monk.lua
@@ -9,7 +9,7 @@ local function useComboBreakerBlackoutKick()
         116768, -- Combo Breaker: Blackout Kick (buff)
         "aura",
         {
-            overlay = { texture = "monk_tiger", position = "Left + Right (Flipped)" },
+            overlay = { texture = "monk_tiger", position = "Right (Flipped)" },
         }
     );
 end
@@ -21,7 +21,20 @@ local function useComboBreakerTigerPalm()
         118864, -- Combo Breaker: Tiger Palm
         "aura",
         {
-            overlay = { texture = "monk_tiger", position = "Left + Right (Flipped)" },
+            overlay = { texture = "monk_tiger", position = "Left", color = { 0, 255, 127 } },
+        }
+    );
+end
+
+local function useVitalMists()
+    SAO:CreateEffect(
+        "vital_mists",
+        SAO.MOP,
+        118674, -- Vital Mists (buff)
+        "aura",
+        {
+            aka = 122107, -- Vital Mists
+            overlay = { stacks = 5, texture = "monk_serpent", position = "Left + Right (Flipped)" },
         }
     );
 end
@@ -29,6 +42,8 @@ end
 local function registerClass(self)
     useComboBreakerBlackoutKick();
     useComboBreakerTigerPalm();
+
+    useVitalMists();
 end
 
 SAO.Class["MONK"] = {